Fix price calculation in Nueva component to use parseFloat instead of parseInt

+++ src/Pages/Cotizaciones/Nueva.jsx
@@ -48,7 +48,7 @@ export default function Nueva() {
                 // lanzarError = true;
                 return;
             }
-            sub += parseInt(s.price);
+            sub += parseFloat(s.price);
         });
 
         // if(lanzarError)
